Add Number interface, implement with Secret and Guess. Refactor Diff with test
To unit test Diff (test in isolation without involving Secret and Guess code), we need to use either a stub or a fake. Because Secret and Guess are final, to stub or fake these we need to implement an interface. Both have a single method number(), so let's have an interface for Number, stub that, and then Diff becomes an object that is about finding the difference between 2 numbers -- reusable beyond just Secret and Guess! Bonus!